Moved types into <types.h> header
