+++ src/lib/include/lval.h
@@ -2,29 +2,8 @@
 #define CURLY_LVAL_H
 
 #include <builtins.h>
-
-/// \brief Represents a Lisp Value 
-/// 
-/// A `lval` consists of a:
-/// - type      : int corresponding to an enum value 
-/// - num       : long coresonding to a number
-/// - err       : char* corresponding to an error message (optional)
-/// - sym       : char* corresponding to a symbol or operator (optional)
-/// - count     : int corresponding to the number of elements in the `cell` array
-/// - cell      : lval** corresponding to an array of lvals
-struct lval;
-typedef struct lval lval;
-
-
-/// \brief Enum for possible lval types
-///
-/// The possible lval types are:
-/// - LVAL_ERR : Error type
-/// - LVAL_NUM : Number type
-/// - LVAL_SYM : Symbol type
-/// - LVAL_SEXPR : S-Expression type
-enum { LVAL_ERR, LVAL_NUM, LVAL_SYM, 
-       LVAL_FUN, LVAL_SEXPR, LVAL_QEXPR };
+#include <lenv.h>
+#include <types.h>
 
 
 ///////////////////////////
