+++ src/lib/src/utilities.c
@@ -1,4 +1,5 @@
 #include <utilities.h>
+#include <types.h>
 
 char* ltype_name(int t)
 {
